simplify StudyLegend.
It is not necessary to have a method `cleanUp`, in fact that is precisely _not_ the behaviour we want. We don't want the injection to be removed when the component unmounts; there is no code that calls the injection to be reintroduced when it does get mounted again. The injection should be valid throughout the life of the `<SmartCharts />` component, and removed only when `<SmartCharts />` is unmounted (via `ChartStore.destroy()`).